requirements.txt
google-generativeai
streamlit
python-dotenv
mysql-connector-python